+++ src/init.c
@@ -2,13 +2,13 @@
 #include <R.h>
 #include <Rinternals.h>
 
-extern SEXP pack_();
-extern SEXP unpack_();
+extern SEXP marshall_();
+extern SEXP unmarshall_();
 
 static const R_CallMethodDef CEntries[] = {
 
-  {"pack_"  , (DL_FUNC) &pack_  , 1},
-  {"unpack_", (DL_FUNC) &unpack_, 1},
+  {"marshall_"  , (DL_FUNC) &marshall_  , 1},
+  {"unmarshall_", (DL_FUNC) &unmarshall_, 1},
   {NULL , NULL, 0}
 };