+++ src/serialize.c
@@ -10,143 +10,18 @@
 #include <unistd.h>
 
 
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// The data buffer.
-// Needs total length and pos to keep track of how much data it contains
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-typedef struct {
-  R_xlen_t length;
-  R_xlen_t pos;
-  unsigned char *data;
-} buffer;
-
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Initialise an empty buffer to hold 'nbytes'
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-buffer *init_buffer(int nbytes) {
-  buffer *buf = (buffer *)malloc(sizeof(buffer));
-  if (buf == NULL) {
-    error("init_buffer(): cannot malloc buffer");
-  }
-
-  buf->data = (unsigned char *)malloc(nbytes * sizeof(unsigned char));
-  if (buf->data == NULL) {
-    error("init_buffer(): cannot malloc buffer data");
-  }
-
-  buf->length = nbytes;
-  buf->pos = 0;
-
-  return buf;
-}
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Naive buffer expansion - double it every time space runs out
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void expand_buffer(buffer *buf) {
-
-  double factor = 2;
-  buf->length = (R_xlen_t)(factor * buf->length);
-
-  if (buf->length > R_XLEN_T_MAX) {
-    error("Requested buffer expandsion too large: %td\n", buf->length);
-  }
-
-  unsigned char *new_data = (unsigned char *)realloc((void *)buf->data, buf->length);
-
-  if (new_data == NULL) {
-    free(buf->data);
-    free(buf);
-    error("Couldn't expand buffer to reallocate: %td\n", buf->length);
-  }
-
-  buf->data = new_data;
-
-}
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Write a byte into the buffer at the current location.
-// The actual buffer is encapsulated as part of the stream structure, so you
-// have to extract it first
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void write_byte(R_outpstream_t stream, int c) {
-  buffer *buf = (buffer *)stream->data;
-
-  // Expand the buffer if it's out space
-  while (buf->pos >= buf->length) {
-    expand_buffer(buf);
-  }
-
-  buf->data[buf->pos++] = (unsigned char)c;
-}
-
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Write multiple bytes into the buffer at the current location.
-// The actual buffer is encapsulated as part of the stream structure, so you
-// have to extract it first
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void write_bytes(R_outpstream_t stream, void *src, int length) {
-  buffer *buf = (buffer *)stream->data;
-
-  // Expand the buffer if it's out space
-  while (buf->pos + length > buf->length) {
-    expand_buffer(buf);
-  }
-
-  memcpy(buf->data + buf->pos, src, length);
-
-  buf->pos += length;
-}
-
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Read a byte from the serialized stream
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-int read_byte(R_inpstream_t stream) {
-  buffer *buf = (buffer *)stream->data;
-
-  if (buf->pos >= buf->length) {
-    error("read_byte(): overflow");
-  }
-
-  return buf->data[buf->pos++];
-}
-
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Read multiple bytes from the serialized stream
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void read_bytes(R_inpstream_t stream, void *dst, int length) {
-  buffer *buf = (buffer *)stream->data;
-
-  if (buf->pos + length > buf->length) {
-    error("read_bytes(): overflow");
-  }
-
-  memcpy(dst, buf->data + buf->pos, length);
-
-  buf->pos += length;
-}
-
+#include "buffer-dynamic.h"
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Serialize an R object
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP pack_(SEXP robj) {
+SEXP marshall_(SEXP robj) {
 
   // Create the buffer for the serialized representation
   // See also: `expand_buffer()` which re-allocates the memory buffer if
   // it runs out of space
-  buffer *buf = init_buffer(16384);
+  dynamic_buffer_t *buf = init_buffer(16384);
 
   // Create the output stream structure
   struct R_outpstream_st output_stream;
@@ -182,7 +57,7 @@ SEXP pack_(SEXP robj) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Unpack a raw vector to an R object
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP unpack_(SEXP vec_) {
+SEXP unmarshall_(SEXP vec_) {
 
   if (TYPEOF(vec_) != RAWSXP) {
     error("unpack(): Only raw vectors can be unserialized");
@@ -193,7 +68,7 @@ SEXP unpack_(SEXP vec_) {
   R_xlen_t len = XLENGTH(vec_);
 
   // Create a buffer object which points to the raw data
-  buffer *buf = malloc(sizeof(buffer));
+  dynamic_buffer_t *buf = malloc(sizeof(dynamic_buffer_t));
   if (buf == NULL) {
     error("'buf' malloc failed!");
   }