+++ tests/testthat/test-identical-to-serialize.R
@@ -6,14 +6,14 @@ test_that("identical to base::serialize()", {
 
   expect_identical(
     serialize(mtcars, NULL, xdr = FALSE),
-    pack(mtcars)
+    marshall(mtcars)
   )
 
   yy <- serialize(mtcars, NULL, xdr = FALSE)
 
   expect_identical(
     unserialize(yy),
-    unpack(yy)
+    unmarshall(yy)
   )
 
 
@@ -25,14 +25,14 @@ test_that("identical to base::serialize()", {
 
     expect_identical(
       serialize(zz, NULL, xdr = FALSE),
-      pack(zz)
+      marshall(zz)
     )
 
     yy <- serialize(zz, NULL, xdr = FALSE)
 
     expect_identical(
       unserialize(yy),
-      unpack(yy)
+      unmarshall(yy)
     )