+++ NEWS.md
@@ -1,3 +1,8 @@
-# serializer 0.1.0
+# serializer 0.1.1   2020-09-22
+
+* use 'marshall/unmarshall' function names instead of 'pack/unpack'
+* Internal refactor of C code
+
+# serializer 0.1.0   2020-09-17
 
 * Initial release